+++ src/main/java/io/github/Mine4Cut/Mine4Cut_server/api/auth/SignInController.java
@@ -1,8 +1,9 @@
-package io.github.Mine4Cut.Mine4Cut_server.api.user.controller;
+package io.github.Mine4Cut.Mine4Cut_server.api.auth;
 
 import io.github.Mine4Cut.Mine4Cut_server.api.auth.dto.AuthResponse;
 import io.github.Mine4Cut.Mine4Cut_server.api.user.dto.SignInRequest;
-import io.github.Mine4Cut.Mine4Cut_server.service.user.UserService;
+import io.github.Mine4Cut.Mine4Cut_server.service.auth.AuthService;
+import io.github.Mine4Cut.Mine4Cut_server.service.user.dto.SignInDto;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -12,12 +13,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 @RestController
 public class SignInController {
-    private final UserService userService;
+
+    private final AuthService authService;
 
     @PostMapping("/auth/sign-in")
     public AuthResponse signIn(@RequestBody @Valid SignInRequest req) {
-        String accessToken = userService.signIn(req);
-
-        return AuthResponse.of(accessToken);
+        SignInDto signInDto = authService.signIn(req);
+        return AuthResponse.of(signInDto);
     }
 }
