+++ src/main/java/io/github/Mine4Cut/Mine4Cut_server/api/auth/dto/AuthResponse.java
@@ -1,11 +1,14 @@
 package io.github.Mine4Cut.Mine4Cut_server.api.auth.dto;
 
+import io.github.Mine4Cut.Mine4Cut_server.service.user.dto.SignInDto;
+
 public record AuthResponse(
-        String accessToken
+    SignInDto signInDto
 ) {
-    public static AuthResponse of(String accessToken) {
+
+    public static AuthResponse of(SignInDto accessToken) {
         return new AuthResponse(
-                accessToken
+            accessToken
         );
     }
 }
